requirements.txt
pyTelegramBotAPI>=4.9.1
yt-dlp>=2025.10.14
python-dotenv>=1.0.0
requests>=2.31.0
